Bad greptile didn't see this was supposed to be cast as short

+++ AAEmu.Game/Models/Game/Skills/Effects/RecoverExpEffect.cs
@@ -68,7 +68,7 @@ public class RecoverExpEffect : EffectTemplate
 
         // Use labor and recover exp
         // Note we don't use player.ChangeLabor here as that would generate extra exp from labor consumption
-        player.LaborPower -= neededLaborCost;
+        player.LaborPower -= (short)neededLaborCost;
         player.SendPacket(new SCCharacterLaborPowerChangedPacket(-neededLaborCost, 0, 0, 0));
         player.SendPacket(new SCRecoverableExpPacket(player.ObjId, 0, 0, 1));
         player.AddExp(player.RecoverableExp, false);
